Confirm if person truly should be deleted

+++ part2/phonebook/src/components/ListPersons.jsx
@@ -3,13 +3,20 @@ import Button from './Button'
 import personsService from '../services/persons'
 
 const handleDelete = (person, persons, setPersons) => {
-  // delete person from the DB
-  personsService
-    .deletePerson(person.id)
-    .then(response => {
-        // update the state and remove the person from the list
-        setPersons(persons.filter(p => p.id !== person.id))
-    })
+    // use a window.confirm to ask the user if they are sure they want to delete the person
+    if (window.confirm(`Delete ${person.name}?`)) {
+        // delete person from the DB
+        personsService
+            .deletePerson(person.id)
+            .then(response => {
+                // update the state and remove the person from the list
+                setPersons(persons.filter(p => p.id !== person.id))
+            })
+            .catch(error => {
+                console.log(error)
+                alert('error deleting person')
+            })
+    }
 }
